refactor(project): convert create dialog into a wizard
