+++ ${sys:LOG_DIRECTORY}/application.log
@@ -6,3 +6,4 @@
 2023-05-14 21:20:19.694 [DEBUG] Application initialized
 2023-05-14 21:23:16.024 [DEBUG] Application initialized
 2023-05-14 21:27:00.375 [DEBUG] Application initialized
+2023-05-14 22:12:57.628 [DEBUG] Application initialized
