+++ .idea/gradle.xml
@@ -23,6 +23,14 @@
           </set>
         </option>
       </GradleProjectSettings>
+      <GradleProjectSettings>
+        <option name="externalProjectPath" value="$PROJECT_DIR$/feature-project/metadata" />
+        <option name="modules">
+          <set>
+            <option value="$PROJECT_DIR$/feature-project/metadata" />
+          </set>
+        </option>
+      </GradleProjectSettings>
     </option>
   </component>
 </project>
