+++ feature-project/create/src/jvmMain/kotlin/projectcreate/di/Module.kt
@@ -2,18 +2,12 @@ package projectcreate.di
 
 import org.koin.dsl.module
 import projectcreate.ui.dialog.CreateProjectViewModel
+import projectmedatada.di.projectMetadataModule
 
 val projectCreateModule = module {
+    includes(projectMetadataModule)
+
     factory {
-        CreateProjectViewModel(
-            dispatcherProvider = get(),
-            languageRepository = get(),
-            completeLanguage = get(),
-            segmentUseCase = get(),
-            projectRepository = get(),
-            filePairRepository = get(),
-            segmentRepository = get(),
-            notificationCenter = get(),
-        )
+        CreateProjectViewModel()
     }
 }
