+++ feature-project/metadata/gradle.properties
@@ -1,0 +1,5 @@
+kotlin.code.style=official
+
+kotlin.version=1.8.0
+agp.version=7.3.0
+compose.version=1.3.0
