+++ feature-project/metadata/src/jvmMain/kotlin/projectmedatada/ui/components/FilePairSelector.kt
@@ -1,4 +1,4 @@
-package projectcreate.ui.components
+package projectmedatada.ui.components
 
 import androidx.compose.foundation.ExperimentalFoundationApi
 import androidx.compose.foundation.background
